Rename modules and extract AI controller
